requirements.txt
anyio==3.7.1
APScheduler==3.9.1.post1
cachetools==5.2.1
certifi==2023.7.22
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
python-dotenv==1.0.0
python-telegram==0.18.0
python-telegram-bot==20.0a2
pytz==2023.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
telegram-text==0.1.2
tornado==6.3.2
tzlocal==5.0.1
